Demo TAIST 6th REST/mqtt_app/requirements.txt
paho-mqtt
requests
pymongo
python-dotenv